Add express and setup entry point
