+++ package.json
@@ -29,6 +29,8 @@
     "url": "git+ssh://git@bitbucket.org:arijea/easy-agile-database-service.git"
   },
   "dependencies": {
+    "@types/express": "^4.17.13",
+    "express": "^4.18.1",
     "lodash": "^4.17.21",
     "pg": "^8.7.3",
     "typeorm": "^0.3.7"
